my-react-app/functions/ai_recipe_search/requirements.txt
transformers
torch
sentence-transformers
scikit-learn
Flask
Flask-Cors
requests
google-generativeai
google-cloud-secret-manager